requirements.txt
transformers
huggingface_hub
python-dotenv
torch
accelerate==0.26.0